% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpora.R
\name{categories}
\alias{categories}
\title{List data set categories in the corpora package}
\usage{
categories()
}
\value{
Character vector of category names.
}
\description{
List data set categories in the corpora package
}
