% Generated by roxygen2 (4.0.2): do not edit by hand
\name{corpora}
\alias{corpora}
\title{Load a data set from the corpora package}
\usage{
corpora(which)
}
\arguments{
\item{which}{The data set to load, a string. If not given, then
all data sets in the package are listed.}
}
\value{
A data frame containing the data set.
}
\description{
Load a data set from the corpora package
}

